feat: Add Order model for customer orders with associated package
This commit introduces the Order model, which represents an order placed by a customer for a specific package. The model structure includes the following fields:

customer (ForeignKey): References the customer who placed the order using Django's built-in get_user_model() function.
package (ForeignKey): References the chosen package for the order, linking it to the PackageAdToken model.
first_name (CharField): Stores the first name of the customer.
last_name (CharField): Stores the last name of the customer.
email (EmailField): Stores the email address of the customer.
phone (PhoneNumberField): Stores the phone number of the customer.
order_note (TextField): Allows the customer to provide additional notes for the order (optional).

The following fields are filled based on the associated package information:
price (PositiveIntegerField): Stores the price of the package.
discount (BooleanField): Tracks whether a discount is applicable to the package.
discount_price (PositiveIntegerField): Indicates the discount price of the package
token_quantity (PositiveIntegerField): Stores the quantity of tokens included in the package.

Additional fields include:
completed (BooleanField): Tracks the completion status of the order.
transaction (PositiveBigIntegerField): Generates a unique transaction ID for the order using the generate_short_uuid() function.
datetime_ordered (DateTimeField): Records the date and time when the order was placed.
datetime_paid (DateTimeField): Records the date and time when the order was paid.

In addition, a new function generate_short_uuid() is implemented to generate a short unique identifier as an integer value using UUIDv4. This function is utilized to generate the transaction ID for each order. The generated transaction ID ensures uniqueness and serves as a concise identifier for each order.

By incorporating the generate_short_uuid() function, the system can generate distinct transaction IDs for different orders, enhancing traceability and identification of orders within the system.

This commit adds the Order model to facilitate the storage and management of customer orders, capturing relevant details such as customer information, package selection, pricing, and timestamps.