Add MAX_DISCOUNT_PERCENT variable to settings.py
This commit introduces the MAX_DISCOUNT_PERCENT variable to the settings.py file. This variable will be used in the PackageAdTokenForm class to determine the maximum discount percentage allowed for a package's discount price. By adding this variable to the settings file, we ensure that the discount calculation is configurable and can be easily adjusted as per the requirements.

feat: Implement PackageAdTokenForm with enhanced validation, user permissions, and additional features

This commit introduces the new PackageAdTokenForm class. It is a subclass of Django's ModelForm and is used to handle the creation and editing of instances of the PackageAdToken model.

The form includes all fields from the PackageAdToken model (fields = '__all__').

The overridden clean() method performs additional data validation based on specific conditions. It starts by retrieving the current user from the request object.

If an instance is marked as deleted (is_delete=True) or if clean data indicates that it should not be deleted (is_delete=False), further validation steps are executed. Otherwise, appropriate feedback messages are displayed using Django's messaging framework, and clean data is returned as an empty dictionary to prevent any modifications.

Within this validation process:
An attempt is made to calculate the token price by multiplying AD_TOKEN_PRICE (a setting) with the value of token_quantity. If a TypeError occurs during this calculation, indicating invalid input data, clean data is returned without any changes.

If the 'discount_price' field is filled out but the 'discount' field is empty, an error will be displayed. This ensures that if a discount price is provided, the discount field must also be set to True.

Validation of the `MAX_DISCOUNT_PERCENT` setting has been intensified to ensure compliance within the accepted range (0 < max_discount_percent < 100). Should this setting fall outside the valid range, an error message will be shown, prompting users to report it to the backend developer. Furthermore, in such cases, the form resets `clean_data` to an empty dictionary, preventing unintended processing. Then return clean_data.

If the 'discount' field is set to True but no value is provided for 'discount_price', an error message will be displayed indicating that a discount price must be filled out when there's an active discount.

It validates that if both 'price_token' and 'discount_price' are present (not None), then 'discount_price' should be less than 'price_token'; otherwise, it displays an error stating that "The equal of discount price must be less than from price".

If 'discount_price' is smaller than min_discount_price (calculated based on MAX_DISCOUNT_PERCENT(setting)), it throws an appropriate error indicating that "The minimum discount price should not exceed {min_discount_price}".

After validating pricing and discounts:
For new instances (those without primary keys), 'created_by' attribute of that instance will be set with information about the current user who made these changes. The 'price' attribute will also be updated accordingly based on token_quantity calculations.
For existing instances (those with primary keys), if has_change is True, then the 'edited_by' attribute of that instance will be updated with information about the current user who made these changes. Additionally, the 'price' attribute will be updated based on token_quantity calculations.

Superusers have additional permissions within this form:
If clean data indicates a change in confirmation status from True to False, and if it's a new instance, the 'confirmed_by' will set with current superuser in clean_data. And if not the attribute 'confirmed_by' of instance will set with current superuser.

Similarly, if an existing instance has its confirmation status set to True but clean data indicates it should now be False, the 'unconfirmed_by' attribute of that instance is updated with information about the current user.

If an object was marked as deleted (is_delete=True), but clean data shows it should no longer be deleted (is_delete=False), and if the current user is a superuser, then the 'undelete_by' attribute of that instance will be set accordingly.

However, normal users do not have permission to change either confirmation or is_delete. These fields are automatically handled by their respective logic within this form. So confirmation will set False for normal user.

These enhancements provide improved validation and implement appropriate permissions for package creation and editing based on various conditions.