fix: Update datetime_modified Field in PackageAdToken Model
This commit addresses an issue in the `PackageAdToken` model where the `datetime_modified` field was not being automatically updated when the record was modified. The `datetime_modified` field was previously defined with `null=True`, which allowed it to be nullable but did not automatically update the value on modifications.

To resolve this, the `datetime_modified` field has been updated with the `auto_now=True` parameter. This modification enables the field to automatically update with the current timestamp whenever any changes are made to the record.

feat: Customize PackageAdToken ModelAdmin for Efficient Package Management

This commit introduces a custom ModelAdmin, `PackageAdTokenAdmin`, to enhance the administration interface for the `PackageAdToken` model. The new admin interface includes several features to improve package management and user experience.

1. Enhanced Display:
   - The admin list display now includes fields such as `name`, `price`, `discount`, `discount_price`, `token_quantity`, `datetime_modified`, `confirmation`, and `is_delete`.
   - The packages are sorted in descending order based on the `datetime_modified` field.

2. Filtering and Searching:
   - Filters for `is_delete` and `confirmation` statuses are provided for easy package filtering.
   - A search field is available for searching packages based on token quantity, name, and price.

3. Form Customization:
   - A custom `PackageAdTokenForm` is utilized to validate package data, ensuring consistent and accurate input.
   - The `get_form` method in the `PackageAdTokenAdmin` sets `form.request = request`, `form.instance = obj`, and `form.change = change` to provide access to request data, the current object instance, and change status in the form.

4. Read-Only Fields:
   - Certain fields, such as `price`, `datetime_created`, and others, are marked as read-only to prevent unauthorized changes.
   - Non-superusers are restricted from editing the `confirmation` field, and if a package is marked for deletion, relevant actions are disabled.
   - If a package is marked as deleted, the `is_delete` field is added to the displayed fields, allowing administrators to track package deletion status.
   - When the logged-in user is not a superuser, the `is_delete` field is also added to `readonly_fields`, ensuring non-superusers cannot modify the deletion status.

5. Save and Delete Functionality:
   - When creating a new package, the admin automatically tracks the `created_by` and sets the correct `price`.
   - For existing packages, the admin tracks `edited_by` and updates the `price` if any changes are made.
   - The `delete_model` method ensures that packages are logically deleted using the `soft_delete` method.

6. Restrict Actions:
   - The admin interface omits any default actions provided by Django.

These improvements optimize the administration of advertisement token packages, providing a more user-friendly and efficient experience for managing and monitoring package details. The custom admin interface enhances package management, maintains data integrity, and empowers administrators with advanced control over the advertisement token packages. The addition of the `get_form` method ensures seamless communication between the custom form and the admin, enabling the form to access relevant information and make informed decisions during package creation and editing. Non-superusers are appropriately restricted from modifying critical fields like `confirmation` and `is_delete`, ensuring data integrity and preserving the system's overall security.