feat: Implement Custom Manager for Active Package Management
This commit introduces a custom manager, `ActivePackageManager`, to the `PackageAdToken` model. This custom manager refines the querying capabilities of the model by providing a specialized method for retrieving active packages.

The `ActivePackageManager` class extends the functionality of the default manager, enabling the retrieval of packages that meet specific criteria. This custom manager filters packages based on their confirmation status (True) and ensures that they are not marked for deletion (is_delete=False).

Additionally, the `PackageAdToken` model now utilizes this custom manager by assigning it to the `active_objs` attribute. This facilitates streamlined access to active packages that fulfill the predefined conditions.

By implementing this custom manager, the system gains enhanced querying capabilities, allowing for more precise and efficient retrieval of active packages. This improvement contributes to the overall robustness and usability of the PackageAdToken model.

feat: Add Serializer for PackageAdToken Model

This commit introduces a new serializer, PackageAdTokenSerializer, designed to facilitate the serialization of data from the PackageAdToken model.

The serializer defines a Meta class that specifies the model to be serialized and the fields to include in the serialization process.

The fields included in the serializer are 'pk', 'name', 'description', 'price', 'discount', 'discount_price', and 'token_quantity'. These fields are marked as read-only, ensuring that they are not modified through the serialization process.

By adding this serializer, the system gains the capability to efficiently convert PackageAdToken model instances into a structured format, which is essential for data exchange and integration with various components of the application.