feat: Introduce Order Create/Update Serializer
This commit introduces the `OrderCreateOrUpdateSerializer`, a specialized serializer designed for creating or updating `Order` instances. The serializer enhances the data validation and representation process while focusing on essential fields required for creating or updating an order.

Notable features of the `OrderCreateOrUpdateSerializer` include:

1. Phone Number Validation: The `phone` field is configured to utilize the `PhoneNumberField` with the 'IR' (Iran) region, ensuring valid and structured phone number input for Iranian numbers.

2. Simplified Field Set: The serializer is tailored to include only the necessary fields (`package`, `first_name`, `last_name`, `email`, `phone`, `order_note`) for creating or updating an order, streamlining the data handling process.

3. Field Requirements: Crucial fields such as `package`, `first_name`, `last_name`, and `phone` are marked as required and not nullable, contributing to data consistency and integrity.

By introducing the `OrderCreateOrUpdateSerializer`, this commit enhances the validation and handling of order data during creation or updates, promoting accurate and reliable data management within the application.

feat: Implement Detailed Order Read Serializer with Field Customization

This commit introduces the `OrderReadSerializer`, a specialized serializer tailored to provide comprehensive and detailed representations of `Order` instances. The serializer offers a detailed overview of order-related information while customizing the representation of certain fields for enhanced data clarity.

Key features of the `OrderReadSerializer` include:

1. Package Details: The `package` field is serialized using the `PackageAdTokenSerializer`, ensuring a detailed representation of the associated package information.

2. Comprehensive Representation: The serializer includes various fields (`package`, `first_name`, `last_name`, `email`, `phone`, `order_note`, `transaction`, `price`, `discount`, `discount_price`, `token_quantity`) that offer a holistic view of an order's details.

3. Read-Only Fields: The read_only_fields attribute is set to 'fields' to prevent modification of these fields during serialization.

4. Conditional Field Removal: For instances where the `price` is not available (None), the serializer selectively removes fields related to pricing (`price`, `discount`, `discount_price`, `token_quantity`). This feature improves the visual clarity of the serialized data.

By introducing the `OrderReadSerializer`, this commit enhances the serialization process by providing a detailed and customizable representation of order data, contributing to improved data visualization and usability.