feat: Integrate Payment URLs into Main URL Configuration
This commit extends the main URL configuration by integrating the payment.urls module. The URL path 'payment/' has been added to the main urls.py file using the path function with the include method. This inclusion allows the application to route incoming requests related to payments to the appropriate URL patterns defined in the payment.urls module.

By incorporating the payment-related URLs into the main URL configuration, this commit enhances the overall application structure and ensures that payment-related endpoints are correctly directed for processing.

feat: Implement PackageAdToken List APIView

This commit introduces the `PackageAdTokenListAPI`, an APIView that handles the retrieval of a list of active `PackageAdToken` instances.

The APIView utilizes the custom manager `active_objs` to retrieve only active packages and serializes the data using the `PackageAdTokenSerializer` to provide a detailed representation of each package.

Key features of the `PackageAdTokenListAPI` include:

- HTTP GET Endpoint: The APIView responds to HTTP GET requests to retrieve the list of active packages.

- Active Package Filtering: Utilizing the `active_objs` custom manager, the APIView fetches only the active packages, filtering out inactive ones.

- Serialization: The retrieved package data is serialized using the `PackageAdTokenSerializer`, ensuring a comprehensive and structured representation.

- Response: The serialized data is returned as a JSON response with an HTTP status code of 200 (OK), providing the caller with the list of active packages.

By adding the `PackageAdTokenListAPI`, this commit enhances the API functionality by offering a dedicated endpoint for fetching active package information, contributing to improved data accessibility and manipulation.

feat: Add URL Configuration for 'payment' App

This commit introduces the `app_name` variable as 'payment' in the app's URL configuration. The utilization of the `app_name` variable helps namespace the app's URLs, promoting an organized approach to URL naming and management.

Additionally, a new URL pattern is implemented, mapping the path 'package/list/' to the `PackageAdTokenListAPI` view. This view empowers users to access a list of packages within the 'payment' app, thereby enhancing the app's functionality and establishing a structured API endpoint. The URL pattern is named 'packages_list', simplifying reverse URL matching and navigation.

feat: Implement Custom Permission Manager for User Order Ownership

This commit introduces a custom permission manager named IsUserOrderOwner.

This permission manager is designed to verify whether the authenticated user is the owner of a specific order object. The has_object_permission method evaluates if the customer associated with the order matches the requesting user.

By utilizing this custom permission manager, the application ensures that only the rightful owners of orders are granted access to relevant actions and views, enhancing data security and user-specific functionality.