feat: Implement Dynamic Price Calculation in Order Model
This commit introduces the calc_price method to the Order model, enhancing its functionality. The method efficiently determines the appropriate price for an order based on its completion status and associated package properties. If the order is completed (self.price is not None), the method checks for a discount and returns either the discounted price or the original price based on the discount status. If the order is not completed (self.price is None), the method uses the package's discount status to return the corresponding discounted or original package price. This method ensures accurate price calculation while considering the order's completion status and potential future package price changes.

refactor: Sort User Orders by DateTime

This commit sorts the list of user orders in the UserOrdersListAPI view by their datetime_ordered attribute in descending order. This small adjustment improves the user experience by presenting the orders in a more meaningful and chronological manner.

feat: Implement Sandbox Payment Processing via ZarinPal API

This commit introduces the SandBoxProcessPaymentAPI view, responsible for initiating the sandbox payment processing through the ZarinPal API. This view is accessible to authenticated users and follows a series of steps to prepare the payment request.

The view first attempts to retrieve an uncompleted order associated with the authenticated user. If such an order exists, the total price is calculated using the calc_price() method of the order instance. This calculated value is then converted from toman to rial currency. However, if the order associated with the user does not exist, the view responds with a message indicating that the user must register their order first.

The view constructs the ZarinPal request URL, headers, and data, including the merchant ID, amount in rials, description, and callback URL. It sends a POST request to the ZarinPal payment request API and receives a response containing the authority for the payment.

Based on the response, the view generates the payment gateway URL using the obtained authority. If the response contains no errors or empty error messages, the gateway URL is returned to the user for payment processing. Otherwise, any errors are included in the response to provide transparency about any issues encountered during the payment initiation process.

This new API view enhances the application's capabilities by enabling users to initiate and proceed with sandbox payment processing via the ZarinPal gateway, improving the overall user experience when interacting with the payment functionality.

feat: Implement Sandbox Payment Callback Handling via ZarinPal API

This commit introduces the `SandBoxCallPaymentAPI` view, responsible for handling the callback process after a sandbox payment attempt through the ZarinPal gateway. This view is accessible to authenticated users and performs necessary operations to verify the payment status.

The view first attempts to retrieve an uncompleted order associated with the authenticated user. If the order does not exist, the view responds with a message indicating that the user must register their order first.

The view extracts the payment authority and status from the query parameters of the callback URL. It then calculates the total price in toman and rial currency using the `calc_price()` method of the order instance.

If the payment status is 'OK', the view constructs a ZarinPal verification request using the authority, merchant ID, amount in rials, and sends it to the ZarinPal payment verification API. The response from ZarinPal contains the payment status code, which is used to determine the success or failure of the payment.

If the payment is successful (status code 100), the order is closed using the `close_order()` method, which updates the order's fields and increments the user's ad tokens. A response is returned with a 'Done' status and the number of ad tokens bought. In case of a failed payment, the `zarin_errors()` function is used to retrieve the textual representation of the ZarinPal error code, and a response with the error message is returned.

If the payment status is not 'OK', the view responds with a status of 'Failed'.

This new API view enhances the application's functionality by enabling the handling of sandbox payment callbacks through the ZarinPal gateway. It ensures accurate verification of payment statuses and facilitates the user's interaction with the payment processing feature.

feat: Integrate ZarinPal Sandbox Payment URLs

This commit adds two new URL patterns to the application's URL configuration.

The URL path 'checkout/' is associated with the `SandBoxProcessPaymentAPI` view, which facilitates the initiation of the ZarinPal sandbox payment process. Users can use this URL to begin the payment for their orders.

Additionally, the URL path 'callback/' is linked to the `SandBoxCallPaymentAPI` view, responsible for handling callbacks and verifying payment status for completed orders in the ZarinPal sandbox environment.

These URLs enhance the application's payment functionality and allow for seamless payment processing and verification.