refactor: Update 'has_free_ad_quota' Method in CustomUser Model
This commit refines the 'has_free_ad_quota' method within the `CustomUser` model.

The method's functionality remains unchanged, but now it considers only advertisements that have not used an ad token (`is_use_ad_token=False`) when counting the user's ads created within the last 30 days.

This adjustment ensures that the ad quota calculation aligns with the specified conditions for free ads and provides accurate information for ad creation limits.

feat: Implement 'try_using_ad_token' method in CustomUser model

This commit introduces the 'try_using_ad_token' method to the `CustomUser` model.

The method serves the purpose of attempting to use an ad token when the 'can_use' parameter is set to 'True' (str) and the user has available tokens. By decreasing the user's ad_token count by one and saving the changes, the method effectively manages the ad token utilization. If the conditions are met and a token is successfully used, the method returns 'True', indicating the successful deduction. If no token is used or the conditions are not met, the method returns 'False'.

This addition enhances the model's functionality by providing users with the ability to conveniently and accurately manage ad token usage within the application.

refactor: Revise CustomUser and CodeVerify Admin Configuration

In the `admin.py` file of the accounts app, the following changes have been applied:

For the `CustomUserAdmin` class (associated with the `CustomUser` model):

1. Added the 'ad_token' field to the `fieldsets`, extending the user fieldsets to include this new field.

2. Enhanced the `save_form` method to handle the creation of superusers more effectively. When creating a superuser, the 'is_superuser' and 'is_admin' flags are now explicitly set to 'False'.

For the `CodeVerifyAdmin` class (associated with the `CodeVerify` model):

3. Renamed the class from `CustomUserAdmin` to `CodeVerifyAdmin` to align with the model it manages.

These adjustments optimize the user management functionalities in the admin panel, ensuring consistent behavior and improved handling of superuser creation. The changes also enhance the organization of the admin configuration to align with the updated model and requirements.